refactor(aws): Make cloudformation controllers consistent with rest of project
