+++ clouddriver-aws/src/main/groovy/com/netflix/spinnaker/clouddriver/aws/controllers/CloudFormationController.java
@@ -14,34 +14,36 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.clouddriver.controllers;
+package com.netflix.spinnaker.clouddriver.aws.controllers;
 
-import com.netflix.spinnaker.clouddriver.model.CloudFormation;
-import com.netflix.spinnaker.clouddriver.model.CloudFormationProvider;
+import com.netflix.spinnaker.clouddriver.aws.model.CloudFormation;
+import com.netflix.spinnaker.clouddriver.aws.model.CloudFormationProvider;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
-import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.util.AntPathMatcher;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.HandlerMapping;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Slf4j
-@RequestMapping("/cloudFormation")
+@RequestMapping("/aws/cloudFormation/stacks")
 @RestController
 class CloudFormationController {
 
   @Autowired
   private List<CloudFormationProvider> cloudFormationProviders;
 
-  @RequestMapping(method = RequestMethod.GET, value = "/list/{accountId}")
-  List<CloudFormation> list(@PathVariable String accountId,
+  @RequestMapping(method = RequestMethod.GET)
+  List<CloudFormation> list(@RequestParam String accountId,
                             @RequestParam(required = false, defaultValue = "*") String region) {
     log.debug("Cloud formation list stacks for account {}", accountId);
     return cloudFormationProviders
@@ -51,8 +53,10 @@ class CloudFormationController {
       .collect(Collectors.toList());
   }
 
-  @RequestMapping(method = RequestMethod.GET, value = "/get")
-  CloudFormation get(@RequestParam String stackId) {
+  @RequestMapping(method = RequestMethod.GET, value = "/**")
+  CloudFormation get(HttpServletRequest request) {
+    String pattern = (String) request.getAttribute(HandlerMapping.BEST_MATCHING_PATTERN_ATTRIBUTE);
+    String stackId = new AntPathMatcher().extractPathWithinPattern(pattern, request.getRequestURI());
     log.debug("Cloud formation get stack with id {}", stackId);
     return cloudFormationProviders
       .stream()
