+++ clouddriver-aws/src/main/groovy/com/netflix/spinnaker/clouddriver/aws/model/CloudFormation.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.spinnaker.clouddriver.model;
+package com.netflix.spinnaker.clouddriver.aws.model;
 
 import java.util.Date;
 import java.util.List;
